temperature_display/requirements.txt
RPLCD == 1.1.*
smbus-cffi == 0.5.*
paho-mqtt == 1.3.*
